Fix comparison warnings

+++ src/Polynomial.inl
@@ -29,10 +29,10 @@ Polynomial<Order - 1, ParamT>
 
     typename DerivedPolynT::ParamArrT derived_params;
 
-    for (int param_idx = 0; param_idx < Order; ++param_idx)
+    for (unsigned param_idx = 0; param_idx < Order; ++param_idx)
     {
         derived_params[param_idx] =
-            (Order - param_idx) * polyn_.getParam(param_idx);
+            static_cast<float>(Order - param_idx) * polyn_.getParam(param_idx);
     }
 
     return {derived_params};
